surrender

+++ handlers.py
@@ -384,8 +384,9 @@ def query_handler(call):
             enemy = find["_id"]
         tictac.delete_one(find)
         bog.delete_message(int(find["_id"]), int(find["mid"]))
-        bog.delete_message(int(find["_id2"]), int(find["mid2"]))
-        bog.send_message(enemy, f"Игрок @{call.message.chat.username} сдался")
+        if find["mid2"] is not None:
+            bog.send_message(enemy, f"Игрок @{call.message.chat.username} сдался")
+            bog.delete_message(int(find["_id2"]), int(find["mid2"]))
         bog.edit_message_text(
             f"Вы сдались", call.message.chat.id, call.message.message_id
         )
